% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_visits}
\alias{cluster_visits}
\title{Cluster Visits}
\usage{
cluster_visits(visits_vectors, visit_table, spec, cluster_number = 5,
  nstart = 50, iter.max = 15)
}
\arguments{
\item{visits_vectors}{A matrix of embeddings of visits}

\item{visit_table}{A data frame with columns:
\describe{
  \item{\code{visit_id}}{}
  \item{\code{specialties}}{Specialties of doctors, separated by ", "}
}}

\item{spec}{A specialty to be clustered}

\item{cluster_number}{A number of desired clusters (default: 5)}

\item{nstart}{Optionally, \code{nstart} argument of \code{kmeans} (default: 50)}

\item{inter.max}{Optionally, \code{inter.max} argument of \code{kmeans} (default: 15)}
}
\value{
The result of k-means with: cluster, centers etc.
}
\description{
Cluster visits for given specialty of doctors, based on
embeddings of visits. The clustering is performed by
the k-means algorithm.
}
\examples{
inter_term_vectors <- embed_terms(interviews, embedding_size = 10L,
                                  term_count_min = 1L)
exam_term_vectors <- embed_terms(examinations, embedding_size = 10L,
                                 term_count_min = 1L)
visits_vectors <- embed_list_visits(interviews,
                                    examinations, inter_term_vectors, exam_term_vectors)
clusters <- cluster_visits(visits_vectors, visits,
                           spec = "internist",
                           cluster_number = 2L)
clusters$size

sum(clusters$withinss)
clusters$centers

}
