% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_embeddings.R
\name{embed_list_visits}
\alias{embed_list_visits}
\title{Generate Embeddings of Visits from a List}
\usage{
embed_list_visits(inter_descriptions, exam_descriptions,
  inter_term_vectors, exam_term_vectors)
}
\arguments{
\item{inter_descriptions}{A vector of descriptions of interviews, named by visits' IDs;
terms in the descriptions are separated by ", "}

\item{exam_descriptions}{A vector of descriptions of examinations, named by visits' IDs;
the visits' IDs do not have to be identical to inter_descriptions}

\item{inter_term_vectors}{A matrix of embeddings of the interview terms}

\item{exam_term_vectors}{A matrix of embeddings of the examination terms}
}
\value{
A matrix of embeddings of visits -- in each row there is an embedding of one visit.
}
\description{
Generate embeddings of given visits, contained in two named vectors. One
vector contains the descriptions of interviews and the second -- the
descriptions of examinations. This function interatively calls
\link{embed_visit} function. This function on about 100k visits should
take about 4 minutes on i7-7700HQ CPU.
}
\examples{
inter_term_vectors <- embed_terms(interviews, embedding_size = 10L,
  term_count_min = 1L)
exam_term_vectors <- embed_terms(examinations, embedding_size = 10L,
  term_count_min = 1L)

visits_vectors <- embed_list_visits(interviews,
                                    examinations,
inter_term_vectors, exam_term_vectors)
nrow(visits_vectors)
visits_vectors
visits_vectors[is.na(visits_vectors)] <- 0
D <- as.matrix(dist(visits_vectors))

# Two closest visits
examinations[rownames(which(D == min(D[D > 0]),
                  arr.ind = TRUE))]

visits_vectors <- embed_list_visits(interviews, "",
  inter_term_vectors, exam_term_vectors)
nrow(visits_vectors)
visits_vectors

}
