% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_embeddings.R
\name{embed_visit}
\alias{embed_visit}
\title{Generate an Embedding of a Visit}
\usage{
embed_visit(visit_description, inter_term_vectors, exam_term_vectors)
}
\arguments{
\item{visit_description}{A named two-element vector (names: "inter" and "exam")
of concatenated terms of interview and examination. Terms must be separated by ", ".}

\item{inter_term_vectors}{A matrix of embeddings of the interview terms}

\item{exam_term_vectors}{A matrix of embeddings of the examination terms}
}
\value{
A vector of the embedding of the visit.
}
\description{
Generate an embedding of a given visit based on interview and examination descriptions
and the embeddings of terms.
This function calls \link{embed_description} function and merge the obtained embeddings
of the interview and the examination.
}
\examples{
inter_term_vectors <- embed_terms(interviews, embedding_size = 10L,
 term_count_min = 1L)
exam_term_vectors <- embed_terms(examinations, embedding_size = 10L,
  term_count_min = 1L)

embed_visit(c(inter = "", exam = "heart, woman, eye"),
  inter_term_vectors, exam_term_vectors)
embed_visit(c(inter = "cough",
             exam = "heart, woman, eye"),
           inter_term_vectors, exam_term_vectors)
embed_visit(c(inter = "cough", exam = ""),
           inter_term_vectors, exam_term_vectors)
embed_visit(c(inter = "", exam = "objaw chorobowy"),
           inter_term_vectors, exam_term_vectors)

}
\seealso{
\code{\link{embed_description}}, \code{\link{embed_list_visits}}
}
