% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analogies.R
\name{evaluate_term_embeddings}
\alias{evaluate_term_embeddings}
\title{Evaluate Term Embeddings}
\usage{
evaluate_term_embeddings(term_vectors, terms_pairs = NULL, n_max = 5L)
}
\arguments{
\item{term_vectors}{A matrix of embeddings of the terms}

\item{n_max}{A maximum number of neighbors included in the context (default: 5)}

\item{terms_pairs:}{A list of lists of the test pairs, each list of pairs should
contain two vectors of terms; if the list contains the element "synonym", the
\link{synonym_task} is performed for these terms, for the rest pairs there
is performed \link{analogy_task}}
}
\value{
A data frame of accuracies of the analogy task and the synonym task
for each given list of pairs for the context from 1 to \code{n_max}.
}
\description{
Evaluate given term embeddings based on \link{analogy_task} and \link{synonym_task}.
The list of test pairs should be given. The \link{analogy_task} is performed
for each type of given analogies (initially, there are 7 proposed types of analogies,
see \link{terms_pairs_test}). The accuracy is measured for the context size from 1 (the most
restrictive) to the given \code{n_max}.
}
\examples{
term_vectors <- embed_terms(examinations, embedding_size = 5L,
  term_count_min = 1L)
evaluate_term_embeddings(term_vectors, terms_pairs = terms_pairs_test)

}
\seealso{
\code{\link{analogy_task}}, \code{\link{synonym_task}}, \code{\link{terms_pairs_test}}
}
