% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interviews}
\alias{interviews}
\title{Descriptions of Interviews with Patients}
\format{A character vector of legth 10 containing the terms separated by \code{", "}.}
\usage{
interviews
}
\description{
An artificial dataset containing terms from generated
descriptions of interviews with patients.
(cf. \link{examinations} and \link{recommendations}).
}
\details{
The structure of the visit, i.e. the partition to interview, examination
and recommendations is based on real data collected from Polish
health centers. The procedure of extracting medical terms from the real
medical data is described in the paper Dobrakowski et al., 2019.
}
\references{
Dobrakowski, A., A. Mykowiecka, M. Marciniak, W. Jaworski, and P. Biecek 2019.
Interpretable Segmentation of Medical Free-Text
Records Based on Word Embeddings. arXiv preprint arXiv:1907.04152.
}
\keyword{datasets}
