% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{terms_categories}
\alias{terms_categories}
\title{Categories of Medical Terms}
\format{A data frame with 19 rows and 2 variables:
\describe{
  \item{term}{A medical term}
  \item{category}{A category of the term}
}}
\usage{
terms_categories
}
\description{
A dataset containing example terms that could be
extracted by linguists from the medical free-text records.
The procedure of extracting medical terms from the real
medical data is described in the paper Dobrakowski et al., 2019.
Each term has been assigned to at least one category.
}
\references{
Dobrakowski, A., A. Mykowiecka, M. Marciniak, W. Jaworski, and P. Biecek 2019.
Interpretable Segmentation of Medical Free-Text
Records Based on Word Embeddings. arXiv preprint arXiv:1907.04152.
}
\keyword{datasets}
