% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{terms_pairs_test}
\alias{terms_pairs_test}
\title{Test Pairs of Terms to Evaluate Embedding Quality}
\format{A list with elements:
\describe{
  \item{spec}{Terms with relation: specialty -- body part (e.g. "cardiologist" -- "heart");
    a list of two character vectors of length 2}
  \item{person}{Terms with relation: man -- woman;
    a list of two character vectors of length 2}
  \item{synonym}{Synonym terms;
    a list of two character vector of length 1}
}}
\usage{
terms_pairs_test
}
\description{
An example dataset containing pairs of analogies of medical terms to evaluate the quality
of embedddings of terms by functions \link{analogy_task} and \link{synonym_task}.
}
\details{
The assessing the quality of the real-data embeddings was performed by 7 types of analogies, described
in the paper Dobrakowski et al., 2019.
}
\references{
Dobrakowski, A., A. Mykowiecka, M. Marciniak, W. Jaworski, and P. Biecek 2019.
Interpretable Segmentation of Medical Free-Text
Records Based on Word Embeddings. arXiv preprint arXiv:1907.04152.
}
\keyword{datasets}
