% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{visits}
\alias{visits}
\title{Medical Visits from Health Centers}
\format{A data frame with 11 rows and 4 variables:
\describe{
  \item{visit_id}{ID of the visit}
  \item{icd10}{ICD-10 code of the visit}
  \item{doctor_id}{ID of the doctor who led the visit}
  \item{specialties}{all specialties of the doctor, separated by ", "}
}}
\usage{
visits
}
\description{
An example artificial dataset containing
the basic information (ICD-10 code, doctor's ID and
the specialties of the doctors) of 11
medical visits. The dataset is artificial,
but it reflects a structure from a real data collected from Polish health centers.
}
\keyword{datasets}
