% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analogies.R
\name{visualize_analogies}
\alias{visualize_analogies}
\title{Visualize Term Analogies}
\usage{
visualize_analogies(term_vectors, terms, find_analogies = FALSE,
  n = 5L, save = FALSE, path_to_save)
}
\arguments{
\item{term_vectors}{A matrix of embeddings of terms}

\item{terms}{A list of two character vectors: the first and the second elements of analogy pairs}

\item{find_analogies}{A logical indicating if the term analogy task should be performed
before the plotting (default: FALSE)}

\item{n}{An optional number of neighbors included in analogies (default: 5),
needed only if \code{find_analogies = TRUE}}

\item{save}{A logical indicating if the plot should be saved to the file}

\item{path_to_save}{An optional string of the path to the target PDF file}
}
\value{
A generated plot of embeddings.
}
\description{
Visualize analogies for given terms. A visualization of embeddings is
two main components from PCA plotted by ggplot on the 2D plane.
Additionally, the terms from each given pair are
connected to one another. If the embeddings are of good quality, the connection
lines should be almost parallel to each other. If \code{find_analogies = TRUE}
only the pairs with at least one solved analogy task are plotted.
}
\examples{
term_vectors <- embed_terms(examinations, embedding_size = 5L,
  term_count_min = 1L)
visualize_analogies(term_vectors, terms_pairs_test$person)
visualize_analogies(term_vectors, terms_pairs_test$person,
  find_analogies = TRUE, n = 10)

}
\seealso{
\code{\link{analogy_task}}, \code{\link{evaluate_term_embeddings}}, \code{\link{terms_pairs_test}}
}
