% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{visualize_icd10}
\alias{visualize_icd10}
\title{Generate and Visualize Embeddings of ICD-10 Codes}
\usage{
visualize_icd10(visits_vectors, visit_table, method = "tsne",
  save = FALSE, path_to_save)
}
\arguments{
\item{visit_table}{A data frame with columns:
\describe{
  \item{\code{visit_id}}{}
  \item{\code{icd10}}{ICD-10 code of the visit}
}}

\item{method}{One of "tsne" (default) or "pca" - a method of generating the plot}

\item{save}{A logical indicating if the plot should be saved to the file}

\item{path_to_save}{An optional string of the path to the target PDF file}

\item{visit_vectors}{A matrix of embeddings of visits}
}
\value{
A generated plot of embeddings.
}
\description{
Generate and visualize embeddings of ICD-10 codes.
An embedding of the code is the simple average of embeddings
of all visits assigned by this code.
The visualization is generated on the 2D plane by t-SNE algorithm
and plotted with ggplot.
If t-SNE cannot be generated (because for example
perplexity is too large for the number of samples), there are plotted
two main components from PCA.
The plot can be optionally saved to the given PDF file.
}
\examples{
inter_term_vectors <- embed_terms(interviews, embedding_size = 10L,
                                  term_count_min = 1L)
exam_term_vectors <- embed_terms(examinations, embedding_size = 10L,
                                 term_count_min = 1L)
visits_vectors <- embed_list_visits(interviews,
                                    examinations, inter_term_vectors, exam_term_vectors)

visualize_icd10(visits_vectors, visits)

}
