% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{visualize_term_embeddings}
\alias{visualize_term_embeddings}
\title{Visualize Embeddings of Terms}
\usage{
visualize_term_embeddings(term_table, term_vectors, category,
  method = "tsne", save = FALSE, path_to_save)
}
\arguments{
\item{term_table}{A data frame with columns:
\describe{
  \item{\code{term}}{A medical term (elements in this column do not have to be unique)}
  \item{\code{category}}{A category of the term}
}}

\item{term_vectors}{A matrix of embeddings of the terms}

\item{category}{A character vector of categories of the terms to be visualized}

\item{method}{One of "tsne" (default) or "pca" - a method of generating the plot}

\item{save}{A logical indicating if the plot should be saved to the file}

\item{path_to_save}{An optional string of the path to the target PDF file}
}
\value{
A generated plot of embeddings.
}
\description{
Visualize given embeddings of terms of specified categories.
The visualization is generated on the 2D plane by t-SNE algorithm
or by PCA (two main components) and plotted with ggplot.
If t-SNE cannot be generated (because for example
perplexity is too large for the number of samples), there are plotted
two main components from PCA.
The plot can be optionally saved to the given PDF file.
}
\examples{
inter_term_vectors <- embed_terms(interviews, embedding_size = 10L,
  term_count_min = 1L)

visualize_term_embeddings(terms_categories, inter_term_vectors,
                          unique(as.character(terms_categories$category)))
visualize_term_embeddings(terms_categories, inter_term_vectors,
                          c("anatomic"), method = "pca")

}
