% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{visualize_visit_embeddings}
\alias{visualize_visit_embeddings}
\title{Visualize Embeddings of Visits}
\usage{
visualize_visit_embeddings(visits_vectors, visit_table, method = "tsne",
  color_by, spec = NULL, clusters, n_doctors = 5L, n_codes = 9L,
  save = FALSE, path_to_save)
}
\arguments{
\item{visit_table}{A data frame with columns:
\describe{
  \item{\code{visit_id}}{}
  \item{\code{icd10}}{(Optional, if \code{color_by == "icd10"}): ICD-10 code of the visit}
  \item{\code{doctor_id}}{(Optional, if \code{color_by == "doctor"}): doctor's ID}
  \item{\code{specialties}}{Specialties of doctors, separated by \code{", "}}
}}

\item{method}{One of \code{"tsne"} (default) or \code{"pca"} - a method of generating the plot}

\item{color_by}{A string, one of \code{{"doctor", "icd10", "cluster"}}}

\item{spec}{A specialty to be plotted (optional, required if \code{color_by != "clusters"})}

\item{clusters}{An output of the function \link{cluster_visits}
(optional, required if \code{color_by == "clusters"})}

\item{n_doctors}{A number of doctors to be colored, the rest will be grey
(optional, required if \code{color_by == "doctor"}), default: 5}

\item{n_codes}{A number of codes to be colored, the rest will be grey
(optional, required if \code{color_by == "icd10"}), default: 9}

\item{save}{A logical indicating if the plot should be saved to the file}

\item{path_to_save}{An optional string of the path to the target PDF file}

\item{visit_vectors}{A matrix of embeddings of visits}
}
\value{
A generated plot of embeddings.
}
\description{
Visualize given embeddings of visits of specified specialties of doctors.
The visualization is generated on the 2D plane by t-SNE algorithm
or by PCA (two main components) and plotted with ggplot.
If t-SNE cannot be generated (because for example
perplexity is too large for the number of samples), there are plotted
two main components from PCA.
Points representing the visits can be colored by doctor ID,
the first letter of ICD-10 code or by cluster number.
The plot can be optionally saved to the given PDF file.
}
\examples{
inter_term_vectors <- embed_terms(interviews, embedding_size = 10L,
                                  term_count_min = 1L)
exam_term_vectors <- embed_terms(examinations, embedding_size = 10L,
                                 term_count_min = 1L)
visits_vectors <- embed_list_visits(interviews,
                                    examinations, inter_term_vectors,
                                    exam_term_vectors)

visualize_visit_embeddings(visits_vectors, visits, "tsne",
                           color_by = "doctor", spec = "internist", n_doctors = 10)
visualize_visit_embeddings(visits_vectors, visits, "pca",
                           color_by = "doctor", spec = "internist", n_doctors = 9)
visualize_visit_embeddings(visits_vectors, visits,
                           color_by = "icd10", spec = "internist", n_codes = 4)

clusters <- cluster_visits(visits_vectors, visits,
                           spec = "internist", cluster_number = 2)
visualize_visit_embeddings(visits_vectors, visits,
                           color_by = "cluster",
                           clusters = clusters)

}
